# openapi.yaml
...
openapi: 3.0.0
info:
  title: service-workbench-on-aws
  description: API collection of service-workbench-on-aws
  contact:
    email: aws-chamonix-dev@amazon.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: v1.4.3
...
paths:
...
  /api/key-pairs:
    get:
      tags:
        - KeyPairs
      summary: Get all key pairs
      operationId: getKeyPairs
      responses:
        '200':
          description: 'Return all key pairs user has access to, including public key but not private key'
    post:
      tags:
        - KeyPairs
      summary: Create a key pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyPair'
      operationId: createKeyPair
      responses:
        '200':
          description: 'Return created key pair, including private key'
  /api/key-pairs/{keyPairId}:
    get:
      tags:
        - KeyPairs
      summary: Get all key pairs
      parameters:
        - name: keyPairId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/KeyPairId'
      operationId: getKeyPair
      responses:
        '200':
          description: 'Return requested keypair including public key but not private key'
    put:
      tags:
        - KeyPairs
      summary: Edit key pair
      parameters:
        - name: keyPairId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/KeyPairId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditKeyPair'
      operationId: editKeyPair
      responses:
        '200':
          description: 'Return requested keypair including public key but not private key'
    delete:
      tags:
        - KeyPairs
      summary: Delete keypair
      parameters:
        - name: keyPairId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/KeyPairId'
      operationId: deleteKeyPair
      responses:
        '200':
          description: 'Return empty object'
  /api/key-pairs/{keyPairId}/deactivate:
    put:
      tags:
        - KeyPairs
      summary: Deactivate key pair
      parameters:
        - name: keyPairId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/KeyPairId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyPairRev'
      operationId: deactivateKeyPair
      responses:
        '200':
          description: 'Return deactivated keypair including public key but not private key'
  /api/key-pairs/{keyPairId}/activate:
    put:
      tags:
        - KeyPairs
      summary: Activate key pair
      parameters:
        - name: keyPairId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/KeyPairId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyPairRev'
      operationId: ActivateKeyPair
      responses:
        '200':
          description: 'Return activated keypair including public key but not private key'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KeyPairId:
      type: string
      minLength: 1
      maxLength: 1024
      pattern: "^[A-Za-z0-9-_]+$"
      example: 'bdf175e3-2966-456b-bd6a-a6125fa33b8c'
    KeyPairRev:
      type: object
      example: { "rev": 0}
...
    CreateKeyPair:
      type: object
      properties:
        name:
          type: string
        desc:
          type: string
      example: {"name": "foo2-key", "desc": "Example key being created"}
    EditKeyPair:
      type: object
      properties:
        rev:
          type: number
        name:
          type: string
        desc:
          type: string
      example: { "rev": 0, "name": "foo2-key", "desc": "Example key being created" }
...
security:
  - bearerAuth: []